@@ -13,9 +13,9 @@
   },
   "dependencies": {
     "@babel/preset-env": "^7.14.8",
-    "@react-navigation/bottom-tabs": "^6.0.3",
+    "@react-navigation/bottom-tabs": "^6.0.5",
     "@react-navigation/native": "^6.0.2",
-    "@react-navigation/native-stack": "^6.0.3",
+    "@react-navigation/native-stack": "^6.1.0",
     "react": "17.0.1",
     "react-native": "0.64.1",
     "react-native-safe-area-context": "^3.3.0",
@@ -37,7 +37,7 @@
     "lint-staged": "^11.1.1",
     "metro-react-native-babel-preset": "^0.64.0",
     "react-test-renderer": "17.0.1",
-    "typescript": "^3.8.3"
+    "typescript": "^4.3.2"
   },
   "resolutions": {
     "@types/react": "^17"
